One rule per file; helpers to `RuleUtils`, model constants to `OptimizerDefaults`; strategies become per-type singletons compared by pointer; rename `EnforcerAxis` to `EnforcedProperty`
