+++ src/Processors/QueryPlan/Optimizations/Cascades/GroupExpression.h
@@ -18,10 +18,10 @@ class IOptimizationRule;
 using OptimizationRulePtr = std::shared_ptr<const IOptimizationRule>;
 
 /// Which physical property a self-referential enforcer adds; `None` for ordinary expressions. Lets
-/// the input resolver stop an enforcer from satisfying its own input by over-providing on a wildcard
-/// axis (empty sort / empty distribution columns), which would form a cycle. A sorted gather is a
+/// the input resolver stop an enforcer from satisfying its own input by over-providing on a
+/// wildcard (empty sort / empty distribution columns), which would form a cycle. A sorted gather is a
 /// `Distribution` enforcer even though it preserves sorting.
-enum class EnforcerAxis : uint8_t
+enum class EnforcedProperty : uint8_t
 {
     None,
     Sorting,
@@ -45,7 +45,7 @@ public:
         , strategy(other_.strategy)
         , description_suffix(other_.description_suffix)
         , inputs(other_.inputs)
-        , enforcer_axis(other_.enforcer_axis)
+        , enforced_property(other_.enforced_property)
     {}
 
     String getName() const;
@@ -78,8 +78,8 @@ public:
 
     ExpressionProperties properties;
 
-    /// Non-`None` for self-referential enforcer expressions (see `EnforcerAxis`).
-    EnforcerAxis enforcer_axis = EnforcerAxis::None;
+    /// Non-`None` for self-referential enforcer expressions (see `EnforcedProperty`).
+    EnforcedProperty enforced_property = EnforcedProperty::None;
 
     std::unordered_set<RulePropertiesKey, RulePropertiesKeyHash> applied_rules;
