+++ src/Processors/QueryPlan/Optimizations/Cascades/Rule.cpp
@@ -1,4 +1,5 @@
 #include <Processors/QueryPlan/Optimizations/Cascades/Rule.h>
+#include <Processors/QueryPlan/Optimizations/Cascades/RuleUtils.h>
 #include <Processors/QueryPlan/BuildRuntimeFilterStep.h>
 #include <Processors/QueryPlan/FilterStep.h>
 #include <Processors/QueryPlan/ExpressionStep.h>
@@ -8,34 +9,6 @@
 
 namespace DB
 {
-
-bool isTopNSort(const IQueryPlanStep & step)
-{
-    const auto * sorting_step = typeid_cast<const SortingStep *>(&step);
-    return sorting_step != nullptr && sorting_step->getType() == SortingStep::Type::Full && sorting_step->getLimit() > 0;
-}
-
-bool isDistributionPassthroughStep(const IQueryPlanStep & step)
-{
-    return typeid_cast<const ExpressionStep *>(&step) != nullptr
-        || typeid_cast<const FilterStep *>(&step) != nullptr
-        || typeid_cast<const BuildRuntimeFilterStep *>(&step) != nullptr;
-}
-
-GroupExpressionPtr makeEnforcerExpression(
-    const GroupExpressionPtr & source,
-    QueryPlanStepPtr step,
-    ExpressionProperties input_required,
-    ExpressionProperties output_properties,
-    EnforcerAxis axis)
-{
-    auto enforcer_expression = std::make_shared<GroupExpression>(std::move(step));
-    enforcer_expression->group_id = source->group_id;
-    enforcer_expression->inputs.push_back({.group_id = source->group_id, .required_properties = std::move(input_required)});
-    enforcer_expression->properties = std::move(output_properties);
-    enforcer_expression->enforcer_axis = axis;
-    return enforcer_expression;
-}
 
 GroupExpressionPtr IOptimizationRule::addTwoStageSplit(Memo & memo, const GroupExpressionPtr & source_expression,
     GroupExpressionPtr partial_expression, QueryPlanStepPtr final_step,
